réorganisation de l'architecture du code + modification du comportement du carousel
