+++ frontend/public/javascript/carousel.js
@@ -1,4 +1,4 @@
-class Caroussel {
+export class Caroussel {
     constructor() {
         this.translateLeft = 0;
         this.translateRight = 0;
@@ -8,7 +8,10 @@ class Caroussel {
     moveLeft = (e) => {
         e.preventDefault();
         this.count--
+        document.querySelector(".right").style.display = "none"
+        document.querySelector(".container-of-the-cards").classList.add("remove-space-left")
         if(this.direction == null || this.direction == 'right') {
+            
             this.direction = 'left';
             this.translateRight = this.translateLeft - (document.querySelector(".card-container").firstElementChild.clientWidth + 5);
             document.querySelectorAll(".card").forEach(element => {
@@ -27,11 +30,15 @@ class Caroussel {
             document.querySelector(".card-container").appendChild(p_prime);
             firstCard.remove();
             document.querySelector(".card-container").style.transform = "translate(" + -this.translateRight + "px)"
+            document.querySelector(".right").style.display = "block"
+            document.querySelector(".container-of-the-cards").classList.remove("remove-space-left")
         })
     }
     moveRight = (e) => {
         e.preventDefault();
         this.count++
+        document.querySelector(".left").style.display = "none"
+        document.querySelector(".container-of-the-cards").classList.add("remove-space-right")
         let lastCard = document.querySelector(".card-container").lastChild;
         let p_prime = lastCard.cloneNode(true);
         let new_pos = (document.querySelector(".card-container").firstElementChild.clientWidth + 5) * this.count
@@ -45,6 +52,8 @@ class Caroussel {
             });
             document.querySelector(".card-container").addEventListener("transitionend", () => {
                 lastCard.remove();
+                document.querySelector(".left").style.display = "block"
+                document.querySelector(".container-of-the-cards").classList.remove("remove-space-right")
             })
         }
         else {
@@ -54,12 +63,19 @@ class Caroussel {
             });
             document.querySelector(".card-container").lastChild.addEventListener("transitionend", () => {
                 lastCard.remove();
+                document.querySelector(".left").style.display = "block"
+                document.querySelector(".container-of-the-cards").classList.remove("remove-space-right")
             })
         }
     }  
+
+    reinit = () => {
+        document.querySelectorAll(".card").forEach(element => {
+            element.style.transform = "none"
+        });
+        document.querySelector(".card-container").style.transform = "none"
+    }
 }
 
-let carousel = new Caroussel();
-document.querySelector(".left").addEventListener("click", carousel.moveRight)
-document.querySelector(".right").addEventListener("click", carousel.moveLeft)
+
